refactor: extract elements of demo page to separate components
align custom form and demo widget to share one row on 1280px breakpoint and up
add qr code placeholder under the form
