+++ src/components/demo/GoBack.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+
+import { useHistory } from "@docusaurus/router";
+
+const GoBack = () => {
+  const router = useHistory();
+
+  return (
+    <div
+      onClick={() => router.push("/")}
+      className="flex items-center gap-1 mt-10 cursor-pointer"
+    >
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        viewBox="0 0 24 24"
+        strokeWidth={1.5}
+        stroke="currentColor"
+        className="w-6 h-6"
+      >
+        <path
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          d="M10.5 19.5L3 12m0 0l7.5-7.5M3 12h18"
+        />
+      </svg>
+      <span>Go Back</span>
+    </div>
+  );
+};
+export default GoBack;
